requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-json-logger==3.2.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-logging==0.50b0
